requirements.txt
anyio==4.9.0
APScheduler==3.10.4
asyncio==3.4.3
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
loguru==0.7.3
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
win32_setctime==1.2.0
